requirements.txt
streamlit
PyPDF2==3.0.1
langchain==0.1.20
langchain-google-genai==1.0.3
faiss-cpu==1.7.4
python-dotenv
google-generativeai
